Core/Cleanup: move repeating code for locale strings retrieval into separate method. Related code cleanup.
--HG--
branch : trunk

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -416,7 +416,7 @@ void ObjectMgr::AddLocaleString(std::string& s, LocaleConstant locale, StringVec
         int idx = GetOrNewIndexForLocale(locale);
         if (idx >= 0)
         {
-            if (int(data.size()) <= idx)
+            if (data.size() <= size_t(idx))
                 data.resize(idx + 1);
 
             data[idx] = s;
